moved velocity
