+++ Makefile
@@ -1,5 +1,5 @@
 SRCS = ./src
-OBJS = $(wildcard $(SRCS)/*.c) $(wildcard $(SRCS)/*/*.c)
+OBJS = $(wildcard $(SRCS)/*.c) $(wildcard $(SRCS)/*/*.c) $(wildcard $(SRCS)/*/*/*.c)
 CC = gcc
 COMPILER_FLAGS = -w
 LINKER_FLAGS = -lSDL2 -lSDL2_image -lSDL2_ttf
